Speed up playwright browser install in CI
We run playwright against every commit, but this currently involves
a step of installing browsers, which takes almost 30 seconds.

This can be optimised by

1. only installing headless chrome
2. caching .ms-playwright so we don't need to download the same browser
   again if the version of playwright hasn't changed

This is based on the suggestions in this thread:
https://github.com/microsoft/playwright/issues/7249

Some people have pointed out that the system deps installed via
--install-deps may interfere with restoring from cache. To
avoid this I've split out this out into a separate install-deps
step.