Add return_url to ethnicity page links
The ethnicity screen can be accessed from both the appointment show page
and the participant show page. We need to be able to have the back link
(and, once implemented, the redirect after form submit) take the user
back to the appropriate page.

Update the participant details component with a mandatory 'return_url'
argument that we can use to populate the backlink href.