Implement form to record a past mammogram
I've used ModelForm here as the data will be saved to a single model,
and the default behaviour ensures both model-level and form-level
validation is run and handled by form.errors.

There is some additional customisation in the form's clean() and save()
to handle fields that are conditionally shown and required.