Rename id->pk wherever we access or refer to a model primary key
Django has two ways to access the primary key of a model:

- the name of the primary key field (universally 'id' in this codebase)
- the 'pk' method, which returns the primary key regardless of its
  actual column name in the database

We have a mix of pk and id in our codebase which leaves some ambiguity
around which one to use. I've opted to amend our code to use pk
wherever we're referring to the primary key of a model.

The common motivation for using pk is the ability to have primary keys
with column names other than id. For our purposes, this isn't
particularly important. I've gone with pk regardless because:

- It's a common, well-understood convention in Django
- It's useful when reading code to be able to easily distinguish model
  primary keys from other types of id (eg - ids for use in HTML, the ids
  used by our Ethnicity model, etc).